requirements.txt
fastapi==0.104.1
sympy==1.12
numpy==1.24.3
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
openai==1.3.0
pydantic==2.5.0